requirements.txt
asgiref==3.8.1
attrs==25.3.0
backports-zoneinfo==0.2.1
black==24.8.0
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.1.8
dj-database-url==1.0.0
django==4.2.24
django-cors-headers==4.4.0
django-filter==24.3
djangorestframework==3.15.2
dotenv==0.9.9
drf-spectacular==0.28.0
drf-spectacular-sidecar==2025.9.1
gunicorn==23.0.0
idna==3.10
importlib-resources==6.4.5
inflection==0.5.1
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
mypy-extensions==1.1.0
packaging==25.0
pathspec==0.12.1
pkgutil-resolve-name==1.3.10
platformdirs==4.3.6
python-decouple==3.8
python-dotenv==1.0.1
pyyaml==6.0.2
referencing==0.35.1
requests==2.32.4
rpds-py==0.20.1
sqlparse==0.5.3
tomli==2.2.1
typing-extensions==4.13.2
uritemplate==4.1.1
urllib3==2.2.3
whitenoise==6.7.0
zipp==3.20.2
